+++ components/paper-datatable-api/src/paper-datatable-api.js
@@ -39,6 +39,7 @@ class DtPaperDatatableApi {
       page: {
         type: Number,
         notify: true,
+        observer: '_pageChanged',
       },
       /**
        * The current size.
@@ -117,6 +118,10 @@ class DtPaperDatatableApi {
       localeDatePicker: {
         type: Object,
       },
+      frozenHeader: {
+        type: Boolean,
+        value: false,
+      },
       resources: {
         value() {
           return {
@@ -145,13 +150,109 @@ class DtPaperDatatableApi {
       },
       timeoutFilter: Number,
     };
+
+    this.listeners = {
+      'iron-resize': '_resizeHeader',
+    };
   }
 
   get behaviors() {
     return [
       Polymer.AppLocalizeBehavior,
+      Polymer.IronResizableBehavior,
     ];
   }
+
+  /** Frozen Mode **/
+
+  _resizeHeader() {
+    if (this.frozenHeader) {
+      let bodyWidth = this._getTbodyWidths();
+      let headerWidth = this._getTheadWidths();
+
+      if (headerWidth && bodyWidth) {
+        /**
+         * Set all width to auto.
+         */
+        this._resizeAllWidthToAuto(bodyWidth);
+        /**
+         * Resize header width following body width.
+         */
+        this._resizeWidth(bodyWidth, headerWidth, 'header');
+        bodyWidth = this._getTbodyWidths();
+        headerWidth = this._getTheadWidths();
+        /**
+         * Resize body width following header width.
+         */
+        this._resizeWidth(headerWidth, bodyWidth, 'body');
+        bodyWidth = this._getTbodyWidths();
+        headerWidth = this._getTheadWidths();
+        /**
+         * Reajust header width with the new width of body.
+         */
+        this._resizeWidth(bodyWidth, headerWidth, 'header', true);
+      }
+    }
+  }
+
+  _resizeAllWidthToAuto(bodyWidth) {
+    bodyWidth.forEach((bodyTrWidth, index) => {
+      this._resizeTd('auto', index, 'header');
+      this._resizeTd('auto', index, 'body');
+    });
+  }
+
+  _resizeWidth(iterateArray, arrayWidth, type, force = false) {
+    iterateArray.forEach((width, index) => {
+      if (width > arrayWidth[index] || force) {
+        this._resizeTd(width, index, type);
+      }
+    });
+  }
+
+  _getTheadWidths() {
+    const frozenHeaderTable = this.$$('#frozenHeaderTable');
+    if (frozenHeaderTable) {
+      const allTheadTrTh = frozenHeaderTable.querySelectorAll('thead tr th');
+      return Object.keys(allTheadTrTh).map(
+        headerThIndex => allTheadTrTh[headerThIndex].offsetWidth
+      );
+    }
+    return null;
+  }
+
+  _getTbodyWidths() {
+    const table = this.$$('table:not(#frozenHeaderTable)');
+    const tbody = table.querySelector('tbody tr');
+
+    if (tbody) {
+      const allTbodyTd = tbody.querySelectorAll('td');
+      return Object.keys(allTbodyTd).map(bodyTdIndex => allTbodyTd[bodyTdIndex].offsetWidth);
+    }
+    return null;
+  }
+
+  _resizeTd(size, columnIndex, type) {
+    let allTheadTrTh = this.$$('tbody tr').querySelectorAll('td');
+    if (type === 'header') {
+      allTheadTrTh = this.$$('#frozenHeaderTable thead tr').querySelectorAll('th');
+    }
+
+    if (size !== 'auto' && size - 52 !== 0) {
+      const sizeWithoutPad = size - 52;
+      Polymer.dom(allTheadTrTh[columnIndex]).firstElementChild.style.width = `${sizeWithoutPad}px`;
+    } else if (size === 'auto') {
+      Polymer.dom(allTheadTrTh[columnIndex]).firstElementChild.style.width = 'auto';
+    }
+  }
+
+  _handleWrapperScroll(event) {
+    if (this.frozenHeader) {
+      this.$$('#headerWrapper').scrollLeft = event.target.scrollLeft;
+    }
+  }
+
+  /** End of frozen mode **/
 
   attached() {
     this._setColumns();
@@ -163,14 +264,29 @@ class DtPaperDatatableApi {
     return value === targetedValue;
   }
 
-  _generateClass(filters, paginate) {
+  _generateClass(frozenHeader, filters, paginate) {
+    if (frozenHeader && filters && paginate) {
+      return 'filters frozen paginate';
+    }
+    if (frozenHeader && filters) {
+      return 'filters frozen';
+    }
+    if (frozenHeader && paginate) {
+      return 'paginate frozen';
+    }
     if (filters && paginate) {
       return 'paginate filters';
-    } else if (filters) {
+    }
+    if (filters) {
       return 'filters';
     }
-
-    return 'paginate';
+    if (paginate) {
+      return 'paginate';
+    }
+    if (frozenHeader) {
+      return 'frozen';
+    }
+    return '';
   }
 
   _nextPage() {
@@ -206,6 +322,11 @@ class DtPaperDatatableApi {
     this._removeRows();
     this._fillRows(data);
     this._fillColumns();
+    this._resizeHeader();
+  }
+
+  _pageChanged(page, oldPage) {
+    this.oldPage = oldPage;
   }
 
   _removeRows() {
@@ -383,6 +504,7 @@ class DtPaperDatatableApi {
 
       this.set(`toggleColumns.${toggleColumnIndex}.hidden`, !isHidden);
     }
+    this._resizeHeader();
   }
 
   _getThDisplayStyle(hidden) {
@@ -396,7 +518,9 @@ class DtPaperDatatableApi {
   _newSizeIsSelected() {
     const newSize = this.$$('paper-listbox').selected;
     if (newSize) {
-      this.page = 0;
+      if (this.oldPage !== null && this.oldPage !== undefined) {
+        this.page = 0;
+      }
       this.size = newSize;
     }
   }